// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/create-meal-http-request.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MealAiSessionHttpRequest } from './meal-ai-session-http-request';
import { MealItemHttpRequest } from './meal-item-http-request';

export interface CreateMealHttpRequest {
    aiSessions?: Array<MealAiSessionHttpRequest> | null;
    comment?: string | null;
    date?: string;
    imageAssetId?: string | null;
    imageUrl?: string | null;
    isNutritionAutoCalculated?: boolean;
    items?: Array<MealItemHttpRequest>;
    manualAlcohol?: number | null;
    manualCalories?: number | null;
    manualCarbs?: number | null;
    manualFats?: number | null;
    manualFiber?: number | null;
    manualProteins?: number | null;
    mealType?: string | null;
    postMealSatietyLevel?: number;
    preMealSatietyLevel?: number;
}