// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/meal-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MealAiSessionHttpResponse } from './meal-ai-session-http-response';
import { MealItemHttpResponse } from './meal-item-http-response';

export interface MealHttpResponse {
    aiSessions?: Array<MealAiSessionHttpResponse>;
    comment?: string | null;
    date?: string;
    favoriteMealId?: string | null;
    id?: string;
    imageAssetId?: string | null;
    imageUrl?: string | null;
    isFavorite?: boolean;
    isNutritionAutoCalculated?: boolean;
    items?: Array<MealItemHttpResponse>;
    manualAlcohol?: number | null;
    manualCalories?: number | null;
    manualCarbs?: number | null;
    manualFats?: number | null;
    manualFiber?: number | null;
    manualProteins?: number | null;
    mealType?: string | null;
    postMealSatietyLevel?: number;
    preMealSatietyLevel?: number;
    qualityGrade?: string;
    qualityScore?: number;
    totalAlcohol?: number;
    totalCalories?: number;
    totalCarbs?: number;
    totalFats?: number;
    totalFiber?: number;
    totalProteins?: number;
}